+++ transport/oneme/max_wclient.go
@@ -41,6 +41,11 @@ func (c *MaxClient) Connect() error {
 func (c *MaxClient) SetEventCallback(cb func(MaxPacket)) { c.onEvent = cb }
 
 func (c *MaxClient) readLoop() {
+	defer func() {
+		if r := recover(); r != nil {
+			logError("recovered in MaxClient.readLoop: %v", r)
+		}
+	}()
 	for {
 		_, message, err := c.conn.ReadMessage()
 		if err != nil {
@@ -111,6 +116,11 @@ func (c *MaxClient) LoginByToken(token string) error {
 }
 
 func (c *MaxClient) keepalive() {
+	defer func() {
+		if r := recover(); r != nil {
+			logError("recovered in MaxClient.keepalive: %v", r)
+		}
+	}()
 	ticker := time.NewTicker(30 * time.Second)
 	defer ticker.Stop()
 	for range ticker.C {